pyproject.toml
[project]
name = "spanreed"
version = "0.1"
description = ""
authors = [{name = "Marcin Bachry", email = "hegel666@gmail.com"}]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "pydantic",
  "pydantic-settings",
  "walrus",
]

[project.optional-dependencies]
dev = [
  "django",
  "factory-boy",
  "pytest",
  "pytest-cov",
  "pytest-env",
  "pytest-redis",
  "tox-factor",
  "tox-pyenv",
]

[tool.setuptools.packages.find]
include = ["spanreed"]

[project.scripts]
spanreed-consumer = "spanreed.commands:spanreed_consumer"
spanreed-requeue = "spanreed.commands:requeue_dead_letter"

[project.entry-points.pytest11]
spanreed = "spanreed.testing.pytest_plugin"

[tool.pytest.ini_options]
addopts = "-vv -W all --tb=short --log-level=DEBUG -W 'ignore:unclosed:ResourceWarning'"

[tool.pytest_env]
SPANREED_SETTINGS_MODULE = "tests.settings"

[tool.black]
line-length = 120
target-version = ['py311']
skip-string-normalization = true

[tool.ruff]
line-length = 120
target-version = 'py312'

[tool.ruff.lint]
select = [
  'A',   # flake8-builtins
  'B',   # flake8-bugbear
  'C4',  # flake8-comprehensions
  'DTZ', # flake8-datetimez
  'E',   # pycodestyle - error
  'ERA', # eradicate
  'F',   # Pyflakes
  'I',   # isort
  'ICN', # flake8-import-conventions
  'ISC', # flake8-implicit-str-concat
  'UP',  # pyupgrade
  'W',   # pycodestyle - warning
]
fixable = ['I', 'UP']
ignore = [
  'E501', # Line too long
  'B008', # Do not perform function call in argument defaults
  'ERA001', # Found commented-out code
]
